Reimplement ability to read in data files
In a previous update to the clawdata class the =: token was removed,
this unfortunately made the implementation of reading the data files
back in much more difficult so it has been added back in.  Note that
this token is not needed by the Fortran and is only used in Python for
reading data files.