Fix bug in writing out iout_aux variable
Previously if a setrun.py file did not include the
*output_aux_components* variable the clawdata write routine would not
write out the data line.  It also had to be lower case for it to
recognize the 'none' and 'all' options.  This patch checks for the type
of *output_aux_components* and checks via the *lower* string method
function to check which option is used.