Remove extraneous include directives
Remove the duplicate paths from the MODULE_PATH variable that were causing
extraneous include directives on every compile line.  This was accomplished
by using the sort command so there is a chance that something might get mucked
up in the order of the includes but that should not matter for Fortran unless
one is using replacement modules with the same name.  May add a better solution
to this later.