Add optional variable RUNEXE to Makefile.common for the command to run EXE
Normally the Makefile for an application specifies EXE as a filename,
such as xclaw, that serves both as the target for 'make .exe' and also
as the executable command for running the code.

In some case, e.g. if using MPI, the command that needs to be run is
different, e.g. mpiexec with various arguments.  In this case you could set
    RUNEXE = "/path/to/mpiexec -n 4 /path/to/xclaw_mpi"
for example in the application Makefile.  Full absolute paths should be used
since this command is typically executed from a different directory specified by
OUTDIR. Also it should be a string in quotes because it must be passed to
the python code runclaw.py as a single argument.

If RUNEXE is not specified in the Makefile (or as an environment variable)
then RUNEXE is set to EXE by default.